requirements.txt
torch==2.5.1
torchvision==0.19.0
numpy==1.26.4
Pillow==8.3.2
tqdm==4.62.3
tensorboard==2.18.0
scikit-learn==0.24.2
